Refactored code by adding containers, actions, constants and reducers
