+++ src/Root/Root.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import { Route} from 'react-router'
 import About from '../Home/About'
-import WorkoutList from '../features/WorkoutList/Containers/Workout'
+import WorkoutList from '../features/WorkoutList/Containers/Workout.Container'
 /*import SideNav from '../features/SideNavBar/Containers/SideNav'*/
 
 export default class Root extends Component {
