+++ src/features/SideNavBar/Containers/SideNav.Container.js
@@ -1,7 +1,7 @@
-import React from 'react'
+import React, {Component} from 'react'
 import {SideNavBar} from '../StyledComponents/SideNavBar'
 
-export default class SideNav extends React.Component {
+export default class SideNav extends Component {
 
     loadSideNavBar = () => {
         return <SideNavBar / >
