+++ src/features/WorkoutList/WorkoutData.json
@@ -1,0 +1,29 @@
+{
+    "App": "workout",
+    "Details": [
+        {
+            "Date": "05/01/2019",
+            "Activity": "Run",
+            "Duration": "40",
+            "Distance": "3.32",
+            "Calories": "428",
+            "Notes": "Speed at 5.2"
+        },
+        {
+            "Date": "05/07/2019",
+            "Activity": "Push ups, Sit ups, Side Pullups, Crunches, Ball Crunches, Bicycle crunches, Rope crunches, Abdomen machine, Glute Machine, Back, Planks",
+            "Duration": "35",
+            "Distance": "",
+            "Calories": "",
+            "Notes": "30 push-ups" 
+        },
+        {
+            "Date": "05/07/2019",
+            "Activity": "Run",
+            "Duration": "40",
+            "Distance": "3.32",
+            "Calories": "425",
+            "Notes": "Speed at 5.2" 
+        }
+    ]
+}
